+++ inst/examples/ex-resample.R
@@ -1,3 +1,11 @@
-## Jackknife
 x <- rnorm(20)
+
+## Bootstrap
+bootstrap(x, do = mean, n = 100)
+
+## Estimate the 25th and 95th percentiles
+quant <- function(x) { quantile(x, probs = c(0.25, 0.75)) }
+bootstrap(x, n = 100, do = mean, f = quant)
+
+## Jackknife
 jackknife(x, do = mean) # Sample mean